+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 #### 1.9.8-SNAPSHOT
 
+- **[UPDATE]** Updated to Kotlin 1.3.10.
 - **[FEATURE]** (`ktx-graphics`) Added `ShapeRenderer.use` to allow safe omission of the `begin()` and `end()` calls.
 
 #### 1.9.8-b5