+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #### 1.9.8-SNAPSHOT
 
 - **[UPDATE]** Updated to Kotlin 1.3.10.
+- **[UPDATE]** Updated to Kotlin Coroutines 1.0.1.
 - **[FEATURE]** (`ktx-graphics`) Added `ShapeRenderer.use` to allow safe omission of the `begin()` and `end()` calls.
 
 #### 1.9.8-b5