+++ CHANGELOG.md
@@ -1,5 +1,14 @@
-#### 1.9.8-SNAPSHOT
+#### 1.9.9-SNAPSHOT
+- **[UPDATE]** Updated LibGDX to 1.9.9.
+- **[UPDATE]** Updated to Kotlin 1.3.10.
+- **[UPDATE]** Updated to Kotlin Coroutines 1.0.1.
+- **[UPDATE]** Updated VisUI to 1.4.1.
+- **[FEATURE]** (`ktx-actors`) Added `stage` factory method that uses named and default parameters to ease `Stage` creation.
+- **[FEATURE]** (`ktx-graphics`) Added `ShapeRenderer.use` to allow safe omission of the `begin()` and `end()` calls.
 - **[FEATURE]** (`ktx-math`) Added `ImmutableVector2`, an immutable equivalent to `Vector2`.
+
+Known issues:
+- **[BUG]** (`ktx-vis`) Tooltips and tabbed panes in VisUI 1.4.1 are broken when used against LibGDX 1.9.9.
 
 #### 1.9.8-b5